requirements.txt
requests
bs4
psycopg2-binary
apache-airflow-providers-cncf-kubernetes
apache-airflow
apache-airflow-providers-postgres
scipy
pandas